// src/app/core/models/settings/mouza.ts
export class Mouza {
    mouzaID: number;
    mouzaName: string;
    mouzaNameBangla: string;
    sLNo: number;
    isActive: boolean;
    mouzaDetails: string;
    thanaID: number;
    dpzid : number;
    dpzName: string;
    thanaName: string;
    clientThanaID: number | null;
    clientMouzeID: number | null;
}